:art: Split resolver into a separate file
